decrypt en makefile, lagrange usando galois, inverso multiplicaivo galois. CHEQUEAR QUE HACER CON 0

+++ lagrange/lagrange.h
@@ -1,1 +1,5 @@
-double lagrange(int k, double X[], double Y[]);
+#include <stdio.h>
+#include <stdint.h>
+#include "../galois/galois.h"
+
+uint8_t lagrange(int k, uint8_t X[], uint8_t Y[]);
